1. added url for add and edit tc in franchisee module

+++ lib/screens/dashboards/admin/employees/all_employees/add_employees.dart
@@ -595,7 +595,7 @@ class _AddEmployeePageState extends State<AddEmployeePage> {
               ..bankDetails = bankDetailsPath;
 
         final updated = await _isarService.update<RegisteredEmployeeModel>(
-            updateRegisteredEmployee, id! as int);
+            updateRegisteredEmployee, id!);
 
         Logger.warning(':::::$updated');
         if (updated) {
